mori/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
deep-translator==1.11.4
Django==5.1.7
django-allauth==65.4.1
django-cors-headers==4.7.0
django-rest-knox==5.0.2
djangorestframework==3.15.2
drf-yasg==1.21.9
faiss-cpu==1.10.0
filelock==3.17.0
fsspec==2025.3.0
ftfy==6.3.1
huggingface-hub==0.29.3
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
open_clip_torch==2.31.0
packaging==24.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
soupsieve==2.6
sqlparse==0.5.3
sympy==1.13.1
timm==1.0.15
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
deep_translator
psycopg2
psycopg2-binary
